Move a couple of trait implementations closer to the implementors to keep the interface/reused component a bit cleaner.
